backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
pydantic==2.11.1
google-adk==1.2.1
google-cloud-bigquery==3.31.0
google-auth==2.38.0
